requirements.txt
# --- FastAPI app ---
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# --- Google Gemini API ---
google-genai>=1.30.0

# --- Alignment ---
faster-whisper>=1.0.3
numpy>=1.26.0

# --- Tunneling ---
pyngrok>=7.2.0

# --- Dev ---
python-dotenv>=1.0.1
pyyaml>=6.0
